Drf_Auth/requirements.txt
asgiref==3.4.1
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
django-jazzmin==2.4.9
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
importlib-metadata==4.10.1
Markdown==3.3.6
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
zipp==3.7.0
